feat: add slide deck controls
